bpo-43693: Add _PyCode_New(). (gh-26375)
This is an internal-only API that helps us manage the many values used to create a code object.

https://bugs.python.org/issue43693

+++ Include/cpython/code.h
@@ -64,7 +64,7 @@ struct PyCodeObject {
 
     /* These fields are set with computed values on new code objects. */
 
-    Py_ssize_t *co_cell2arg;    /* Maps cell vars which are arguments. */
+    int *co_cell2arg;           /* Maps cell vars which are arguments. */
     // These are redundant but offer some performance benefit.
     int co_nlocalsplus;         /* number of local + cell + free variables */
     int co_nlocals;             /* number of local variables */
